+++ src/arraymancer/laser/primitives/matrix_multiplication/gemm.nim
@@ -228,25 +228,43 @@ proc gemm_strided*[T: SomeNumber and not(uint32|uint64|uint|int)](
         const ukernel = cpu_features.x86_ukernel(T, false)
         apply(ukernel)
 
-    when defined(i386) or defined(amd64):
-      when T is float32:
-        if hasAvx512f(): dispatch(x86_AVX512)
-        elif hasFma3():  dispatch(x86_AVX_FMA)
-        elif hasAvx():   dispatch(x86_AVX)
-        elif hasSse():   dispatch(x86_SSE)
-      elif T is float64:
-        if hasAvx512f(): dispatch(x86_AVX512)
-        elif hasFma3():  dispatch(x86_AVX_FMA)
-        elif hasAvx():   dispatch(x86_AVX)
-        elif hasSse2():  dispatch(x86_SSE2)
-      elif T is int32:
-        if hasAvx512f(): dispatch(x86_AVX512)
-        elif hasAvx2():  dispatch(x86_AVX2)
-        elif hasSse41(): dispatch(x86_SSE4_1)
-        elif hasSse2():  dispatch(x86_SSE2)
-      elif T is int64:
-        if hasAvx512f(): dispatch(x86_AVX512)
-        elif hasSse2():  dispatch(x86_SSE2)
+    # for clarity split AVX512 compilation fully from regular
+    when defined(avx512):
+      when defined(i386) or defined(amd64):
+        when T is float32:
+          if hasAvx512f(): dispatch(x86_AVX512)
+          elif hasFma3():  dispatch(x86_AVX_FMA)
+          elif hasAvx():   dispatch(x86_AVX)
+          elif hasSse():   dispatch(x86_SSE)
+        elif T is float64:
+          if hasAvx512f(): dispatch(x86_AVX512)
+          elif hasFma3():  dispatch(x86_AVX_FMA)
+          elif hasAvx():   dispatch(x86_AVX)
+          elif hasSse2():  dispatch(x86_SSE2)
+        elif T is int32:
+          if hasAvx512f(): dispatch(x86_AVX512)
+          elif hasAvx2():  dispatch(x86_AVX2)
+          elif hasSse41(): dispatch(x86_SSE4_1)
+          elif hasSse2():  dispatch(x86_SSE2)
+        elif T is int64:
+          if hasAvx512f(): dispatch(x86_AVX512)
+          elif hasSse2():  dispatch(x86_SSE2)
+    else:
+      when defined(i386) or defined(amd64):
+        when T is float32:
+          if hasFma3():    dispatch(x86_AVX_FMA)
+          elif hasAvx():   dispatch(x86_AVX)
+          elif hasSse():   dispatch(x86_SSE)
+        elif T is float64:
+          if hasFma3():    dispatch(x86_AVX_FMA)
+          elif hasAvx():   dispatch(x86_AVX)
+          elif hasSse2():  dispatch(x86_SSE2)
+        elif T is int32:
+          if hasAvx2():    dispatch(x86_AVX2)
+          elif hasSse41(): dispatch(x86_SSE4_1)
+          elif hasSse2():  dispatch(x86_SSE2)
+        elif T is int64:
+          if hasSse2():  dispatch(x86_SSE2)
     dispatch(x86_Generic)
 
 proc gemm_strided*[T: uint32|uint64|uint|int](