+++ src/arraymancer/tensor/einsum.nim
@@ -240,6 +240,8 @@ proc checkStatement(stmts: NimNode): StatementKind =
     result = skAuto
   of nnkAsgn:
     result = skAssign
+  of nnkCall: # for generic `[]=` assignment
+    result = skAssign
   else:
     error("`einsum` statement must not be of kind `" & $stmt.kind & "`!")
 
@@ -250,10 +252,18 @@ proc getTensorIdx(tensors: NimNode, tensorArgument: seq[NimNode]): seq[TensorIdx
   ## Returns a sequence of TensorIdx objects, i.e. a tensor ident together with the
   ## associated indices.
   proc extractIdentIdx(n: NimNode, compare: NimNode): TensorIdx =
-    let tC = n[0]
+    var
+      tC: NimNode        # the node of the referring tensor
+      tIdx: seq[NimNode] # the accessed indices
+    case n.kind
+    of nnkBracketExpr: # regular `[]`
+      tC = n[0]
+      tIdx = slice(n, 1 .. ^1)
+    else:
+      error("Invalid branch in `einsum`. Kind " & $n.kind & " not allowed!")
     doAssert $tC == $compare, " was " & $tC & " and "  & $compare
-    let tIdx = slice(n, 1 .. ^1)
     result = TensorIdx(t: tC, idx: tIdx)
+
   case tensors.kind
   of nnkBracketExpr:
     # only a single tensor, probably in an `skAssign` statement
@@ -261,7 +271,7 @@ proc getTensorIdx(tensors: NimNode, tensorArgument: seq[NimNode]): seq[TensorIdx
       "statement of `einsum`, only a single argument may be given!"
     result = @[extractIdentIdx(tensors, tensorArgument[0])]
   of nnkInfix:
-    doAssert tensors[0].eqIdent"*"
+    doAssert tensors[0].eqIdent"*", "Only multiplication allowed in `einsum`"
     if tensors[1].kind == nnkInfix:
       result.add getTensorIdx(tensors[1], tensorArgument)
       result.add getTensorIdx(tensors[2], @[tensorArgument[^1]])
@@ -279,6 +289,8 @@ proc findIdx(tensorSeq: seq[TensorIdx], idx: string): (NimNode, int) =
     let resIdx = find(idxStr, idx)
     if resIdx >= 0:
       result = (tIdx.t, resIdx)
+  if result[0].kind == nnkNilLit:
+    error("Could not find a tensor corresponding to index: " & idx & " in " & $tensorSeq)
 
 proc toDuplicates[T](s: seq[T]): OrderedSet[T] =
   ## creates a set of elements ``only`` consisting of the duplicate elements
@@ -320,7 +332,7 @@ proc replaceRhsByContig(rhs: NimNode): NimNode =
   of nnkInfix:
     for i in 0 ..< result.len:
       case result[i].kind
-      of nnkIdent: discard
+      of nnkIdent, nnkOpenSymChoice: discard
       of nnkInfix:
         result[i] = replaceRhsByContig(result[i])
       of nnkBracketExpr:
@@ -331,6 +343,67 @@ proc replaceRhsByContig(rhs: NimNode): NimNode =
     result[0] = makeContigIdent(result[0])
   else:
     error("Unsupported kind for `einsum` RHS statement " & $result.kind)
+
+proc callToBracket(n: NimNode): NimNode =
+  ## Turns a (possibly nested, containing infix) set of calls using
+  ## `[]` into bracket expressions
+  # we turn AST like the following:
+  # c[i] = x[i, k] * y[i, k]
+  # Call
+  #   OpenSymChoice 11 "[]="
+  #   Ident "c"
+  #   Ident "i"
+  #   Infix
+  #     OpenSymChoice 21 "*"
+  #     Call
+  #       OpenSymChoice 18 "[]"
+  #       Ident "x"
+  #       Ident "i"
+  #       Ident "k"
+  #     Call
+  #       OpenSymChoice 18 "[]"
+  #       Ident "y"
+  #       Ident "i"
+  #       Ident "k"
+  # back into:
+  # Asgn
+  #   BracketExpr
+  #     Ident "c"
+  #     Ident "i"
+  #   Infix
+  #     Ident "*"
+  #     BracketExpr
+  #       Ident "x"
+  #       Ident "i"
+  #       Ident "k"
+  #     BracketExpr
+  #       Ident "y"
+  #       Ident "i"
+  #       Ident "k"
+  # this way we don't have to change any of the main macro logic to handle generic contexts
+  case n.kind
+  of nnkCall:
+    let str = n[0].toStrLit.strVal
+    if str == "[]":                     # bracket expression
+      result = nnkBracketExpr.newTree()
+      for i in 1 ..< n.len:
+        result.add callToBracket(n[i])
+    elif str == "[]=":                  # assignment
+      result = nnkBracketExpr.newTree()
+      for i in 1 ..< n.len - 1:         # first handle arguments to LHS
+        result.add callToBracket(n[i])
+      result = nnkAsgn.newTree(result)
+      result.add callToBracket(n[^1])   # now handle RHS
+  of nnkOpenSymChoice:
+    let str = n.toStrLit.strVal
+    doAssert str == "*"
+    result = ident"*"
+  of nnkSym, nnkIdent: result = n
+  else:
+    if n.len == 0: return n
+    result = newTree(n.kind)
+    for i in 0 ..< n.len:
+      result.add callToBracket(n[i])
 
 proc splitLhsRhs(stmtKind: StatementKind,
                  stmt: NimNode): (NimNode, OrderedSet[string], NimNode) =
@@ -346,18 +419,29 @@ proc splitLhsRhs(stmtKind: StatementKind,
   var idxLHS: OrderedSet[string]
   if stmtKind == skAssign:
     # in case of assign, slice off the infix part
-    rhsStmt = stmt[0][1]
-    lhsStmt = stmt[0][0]
-    case lhsStmt.kind
-    of nnkIdent:
-      # left is an ident, so result supposed to be a scalar. Indidces empty set
-      idxLHS = initOrderedSet[string]()
-    of nnkBracketExpr:
-      idxLHS = toOrderedSet(slice(lhsStmt, 1 .. ^1).mapIt($it))
+    case stmt.kind
+    of nnkStmtList:
+      doAssert stmt.len == 1, "nnkStmtList may only contain a single child"
+      return splitLhsRhs(stmtKind, stmt[0]) # ``recurse`` on child of stmt list
+    of nnkAsgn:
+      rhsStmt = stmt[1]
+      lhsStmt = stmt[0]
+      case lhsStmt.kind
+      of nnkIdent:
+        # left is an ident, so result supposed to be a scalar. Indidces empty set
+        idxLHS = initOrderedSet[string]()
+      of nnkBracketExpr:
+        idxLHS = toOrderedSet(slice(lhsStmt, 1 .. ^1).mapIt($it))
+      else:
+        error("Unsupported kind for `einsum` LHS statement " & $lhsStmt.kind)
+    of nnkCall: # open sym choice in generic context for `[]=`
+      doAssert stmt[0].toStrLit.strVal == "[]="
+      lhsStmt = callToBracket(stmt)
+      return splitLhsRhs(stmtKind, lhsStmt) # ``recurse`` on rewritten AST
     else:
-      error("Unsupported kind for `einsum` LHS statement " & $lhsStmt.kind)
+      error("Unsupported kind for `einsum` LHS statement " & $stmt.kind)
   else:
-    rhsStmt = stmt[0]
+    rhsStmt = callToBracket(stmt[0]) # potentially convert generics
   # now patch `rhsStmt` to use the local contiguous tensors
   rhsStmt = replaceRhsByContig(rhsStmt)
   result = (lhsStmt, idxLhs, rhsStmt)
@@ -507,7 +591,6 @@ proc genContiguous(ts: seq[NimNode], subType: NimNode): (seq[NimNode], NimNode) 
       let `tcIdent` = asContiguous[`subType`](`t`, layout = rowMajor, force = true)
     tsCont.add tcIdent
   result = (tsCont, res)
-  # echo res.treeRepr
 
 macro einsum*(tensors: varargs[typed], stmt: untyped): untyped =
   ## Performs Einstein summation of the given `tensors` defined by the `stmt`.